backend/requirements.txt
django
djangorestframework
pymupdf
langchain
openai
chromadb
python-dotenv
sumy
nltk
